chore: update semantic demo logic
