+++ .dumi/components/TemplateSemanticPreview.tsx
@@ -1,0 +1,81 @@
+import React from 'react';
+import useLocale from '../hooks/useLocale';
+import SemanticPreview from './SemanticPreview';
+
+export const locales = {
+  cn: {
+    popup: '弹出菜单元素',
+  },
+  en: {
+    popup: 'Popup element',
+  },
+};
+
+interface BlockProps {
+  component: React.ComponentType<any>;
+  options?: { value: string; label: string }[];
+  defaultValue?: string;
+  style?: React.CSSProperties;
+  [key: string]: any;
+}
+
+const Block: React.FC<BlockProps> = ({ component: Component, options, defaultValue, ...props }) => {
+  const divRef = React.useRef<HTMLDivElement>(null);
+  return (
+    <div ref={divRef} style={{ position: 'absolute', marginBottom: 80 }}>
+      <Component
+        {...props}
+        open
+        placement="bottomLeft"
+        defaultValue={defaultValue}
+        getPopupContainer={() => divRef.current}
+        options={options}
+        styles={{
+          popup: { zIndex: 1 },
+        }}
+      />
+    </div>
+  );
+};
+
+export interface TemplateSemanticPreviewProps {
+  component: React.ComponentType<any>;
+  defaultValue?: string;
+  options?: { value: string; label: string }[];
+  height?: number;
+  onSearch?: (text: string) => void;
+  placeholder?: string;
+  style?: React.CSSProperties;
+  [key: string]: any;
+}
+
+const TemplateSemanticPreview: React.FC<TemplateSemanticPreviewProps> = ({
+  component,
+  defaultValue,
+  options = [
+    { value: 'aojunhao123', label: 'aojunhao123' },
+    { value: 'thinkasany', label: 'thinkasany' },
+  ],
+  height = 200,
+  style = { width: 200 },
+  ...restProps
+}) => {
+  const [locale] = useLocale(locales);
+
+  return (
+    <SemanticPreview
+      semantics={[{ name: 'popup', desc: locale.popup, version: '5.25.0' }]}
+      height={height}
+    >
+      <Block
+        component={component}
+        defaultValue={defaultValue}
+        options={options}
+        style={style}
+        {...restProps}
+      />
+    </SemanticPreview>
+  );
+};
+
+export default TemplateSemanticPreview;
