+++ components/auto-complete/demo/_semantic.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import { AutoComplete } from 'antd';
 
-import SemanticPreview from '../../../.dumi/components/SemanticPreview';
-import useLocale from '../../../.dumi/hooks/useLocale';
-import { Block, locales } from '../../select/demo/_semantic';
+import TemplateSemanticPreview from '../../../.dumi/components/TemplateSemanticPreview';
 
 const mockVal = (str: string, repeat = 1) => ({
   value: str.repeat(repeat),
@@ -11,7 +9,6 @@ const mockVal = (str: string, repeat = 1) => ({
 });
 
 const App: React.FC = () => {
-  const [locale] = useLocale(locales);
   const [options, setOptions] = React.useState([
     { value: 'aojunhao123', label: 'aojunhao123' },
     { value: 'thinkasany', label: 'thinkasany' },
@@ -21,18 +18,13 @@ const App: React.FC = () => {
     !searchText ? [] : [mockVal(searchText), mockVal(searchText, 2), mockVal(searchText, 3)];
 
   return (
-    <SemanticPreview
-      semantics={[{ name: 'popup', desc: locale.popup, version: '5.25.0' }]}
-      height={200}
-    >
-      <Block
-        component={AutoComplete}
-        style={{ width: 200 }}
-        options={options}
-        onSearch={(text: string) => setOptions(getPanelValue(text))}
-        placeholder="input here"
-      />
-    </SemanticPreview>
+    <TemplateSemanticPreview
+      component={AutoComplete}
+      style={{ width: 200 }}
+      options={options}
+      onSearch={(text: string) => setOptions(getPanelValue(text))}
+      placeholder="input here"
+    />
   );
 };
