+++ components/select/index.zh-CN.md
@@ -165,7 +165,7 @@ return (
 | className | Option 器类名           | string          | -      |      |
 | title     | 选项上的原生 title 提示 | string          | -      |      |
 
-## 语义化DOM（Semantic DOM）
+## 语义化 DOM（Semantic DOM）
 
 <code src="./demo/_semantic.tsx" simplify="true"></code>
